backend/requirements.txt

fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
